+++ engines/gob/goblin/goblin.h
