+++ engines/gob/goblin/goblin_v4.cpp
@@ -26,7 +26,7 @@
  */
 
 #include "gob/gob.h"
-#include "gob/goblin.h"
+#include "gob/goblin/goblin.h"
 #include "gob/global.h"
 #include "gob/mult.h"
 #include "gob/map.h"
