+++ engines/gob/goblin/goblin_v7.cpp
@@ -27,7 +27,7 @@
 
 #include "gob/global.h"
 #include "gob/gob.h"
-#include "gob/goblin.h"
+#include "gob/goblin/goblin.h"
 #include "gob/inter/inter.h"
 #include "gob/map.h"
 #include "gob/mult.h"
